requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
configparser==5.2.0
Flask==2.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
pandas==1.4.2
py-quantaq==1.1.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
